Include select comment groups to be important
Having a hlgroup that I think is important be a bit muted is less
annoying than having an unimportant hlgroup stand out a lot.

Also, bufferline.nvim works some magic on the Comment hlgroup. If that
is green, non-selected buffers show up in green, which looks bad. This
is the faster way to make things look okay.